fix(manifest-replicas): Replicas value was not being set if it didn't already have a value, however users expect that they can omit the value when leveraging the use-source-capacity strategy. Fix is setting the value despite there not already being one. Additionally, if a third party integration creates a manifest via JSON that serializes the replica count as an integer, a casting exception occurs, so this includes an update to play nicely with integers. (#5368)

+++ clouddriver-kubernetes/src/main/java/com/netflix/spinnaker/clouddriver/kubernetes/description/manifest/KubernetesSourceCapacity.java
@@ -21,7 +21,7 @@ import com.netflix.spinnaker.clouddriver.kubernetes.description.KubernetesCoordi
 import com.netflix.spinnaker.clouddriver.kubernetes.security.KubernetesCredentials;
 
 public class KubernetesSourceCapacity {
-  public static Double getSourceCapacity(
+  public static Integer getSourceCapacity(
       KubernetesManifest manifest, KubernetesCredentials credentials) {
     KubernetesManifest currentManifest =
         credentials.get(
